refactor: restructure package to flatten filetree
