+++ src/cala/streaming/composer/__init__.py
@@ -1,0 +1,4 @@
+from .pipe_config import StreamingConfig
+from .runner import Runner
+
+__all__ = ["Runner", "StreamingConfig"]
