+++ src/cala/streaming/core/__init__.py
@@ -1,4 +1,6 @@
 from .exchange import DataExchange
 from .parameters import Parameters
+from .store import BaseStore
+from .transformer_meta import TransformerMeta
 
-__all__ = ["DataExchange", "Parameters"]
+__all__ = ["BaseStore", "Parameters", "TransformerMeta", "DataExchange"]
