+++ src/cala/streaming/core/store.py
