--- src/cala/streaming/core/stores/__init__.py
@@ -1,5 +1,0 @@
-from .base import BaseStore
-from .footprints import FootprintStore
-from .traces import TraceStore
-
-__all__ = ["BaseStore", "FootprintStore", "TraceStore"]
