+++ src/cala/streaming/core/transformer_meta.py
