+++ src/cala/streaming/init/common/__init__.py
@@ -1,0 +1,9 @@
+from .footprints import FootprintsInitializer, FootprintsInitializerParams
+from .traces import TracesInitializer, TracesInitializerParams
+
+__all__ = [
+    "FootprintsInitializer",
+    "FootprintsInitializerParams",
+    "TracesInitializer",
+    "TracesInitializerParams",
+]
