--- src/cala/streaming/initialization/__init__.py
@@ -1,2 +1,0 @@
-from .footprints import FootprintsInitializer, FootprintsInitializerParams
-from .traces import TracesInitializer, TracesInitializerParams
