--- src/cala/streaming/motion_stabilization/__init__.py
@@ -1,3 +1,0 @@
-from .rigid_translation import RigidTranslatorParams, RigidTranslator
-
-__all__ = ["RigidTranslatorParams", "RigidTranslator"]
