+++ src/cala/streaming/preprocess/__init__.py
@@ -2,6 +2,7 @@ from .background_removal import BackgroundEraser, BackgroundEraserParams
 from .denoise import Denoiser, DenoiserParams
 from .downsample import Downsampler, DownsamplerParams
 from .glow_removal import GlowRemover
+from .rigid_stabilization import RigidStabilizer, RigidStabilizerParams
 
 __all__ = [
     "BackgroundEraser",
@@ -11,4 +12,6 @@ __all__ = [
     "Downsampler",
     "DownsamplerParams",
     "GlowRemover",
+    "RigidStabilizerParams",
+    "RigidStabilizer",
 ]
