+++ src/cala/streaming/stores/common/__init__.py
@@ -1,0 +1,4 @@
+from .footprints import FootprintStore
+from .traces import TraceStore
+
+__all__ = ["FootprintStore", "TraceStore"]
