+++ src/cala/streaming/util/buffer.py
