+++ src/cala/streaming/util/io.py
