--- src/cala/streaming/utils/metrics.py
@@ -1,8 +1,0 @@
-class QualityMetrics:
-    """Computes quality metrics"""
-
-    def compute_snr(self, component):
-        pass
-
-    def compute_correlation(self, a, b):
-        pass
