--- src/cala/streaming/utils/statistics.py
@@ -1,5 +1,0 @@
-class SufficientStatistics:
-    """Manages sufficient statistics updates"""
-
-    def update(self, frame, estimates):
-        pass
