--- src/cala/streaming/visualization/display.py
@@ -1,12 +1,0 @@
-class StreamingDisplay:
-    """Handles streaming visualization"""
-
-    def create_display_frame(self, frame, components, residuals):
-        pass
-
-
-class MovieWriter:
-    """Handles movie writing"""
-
-    def write_frame(self, frame):
-        pass
