+++ tests/test_streaming_service/unit/test_preprocess/test_background_removal.py
